Add Azure Bicep infrastructure and update to .NET 7
Introduced Bicep infrastructure scripts for setting up backend resources in Azure, like App Service and Cosmos DB. GiftGivingGenerator.API was migrated to .NET 7 and Entity Framework 7.

Using Bicep scripts will allow for Infrastructure as code practices, providing an easy and repeatable way to build and deploy the necessary cloud resources. The migration to .NET 7 and EF 7 keeps the application up-to-date with the latest frameworks, ensuring support and improved performance. Also, replaced SQL server with Cosmos DB in the DbContext configuration.