"Initial database schema setup using Entity Framework migrations
This commit introduces the initial database schema for the GiftGivingGenerator application. The schema includes tables for Persons, Events, DrawingResults, EventPerson, Exclusions, and GiftWish. All tables have foreign keys and constraints set as per the business logic of the application.
Although previous version was using Cosmos DB as a database, the decision was made to switch to SQL Server and use Entity Framework for code-first database designing and management. Therefore, automated database migrations were created and stored in the file 20231021174634_InitialMigration.cs to better handle schema updates and versioning. The DbContext was updated accordingly and the connection string in Program.cs was changed as well. Additionally, the disappearance of Cosmos DB influenced the change in the infrastructure - in the bicep files - where a new SqlServer and SqlDb resources were created, and the old Cosmos DB was removed."